change a lot of things

+++ PasFinis/Paladin/Paladin.cs
@@ -1,5 +1,7 @@
 ﻿namespace Tp.Characters.Paladin;
 
+
+/*
 public class Paladin : Character
 {
     public Paladin(string name,
@@ -7,9 +9,10 @@ public class Paladin : Character
         int pvActual = 95,
         int physicalAttack = 40,
         int magicalAttack = 40,
-        decimal dodgeChance = 0.05m,
-        decimal paradeChance = 0.1m,
-        decimal spellChance = 0.20m)
+        double dodgeChance = 0.05,
+        double paradeChance = 0.1,
+        double spellChance = 0.20)
         : base(name, pv, pvActual, physicalAttack, magicalAttack, Armors.Armors.Mesh, dodgeChance, paradeChance, spellChance)
     {}
 }
+*/
